converted SelectLang from js to tsx and hooks
